requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.40
PyMySQL==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic[email]==1.10.26
beautifulsoup4==4.12.3
python-multipart==0.0.9
